+++ src/main/java/com/example/sparkle/dtos/inputDto/ProductInputDto.java
@@ -1,11 +1,13 @@
 package com.example.sparkle.dtos.inputDto;
 
+import java.math.BigDecimal;
+
 public class ProductInputDto {
     //    DTO Variables
     public Long id;
     public String productName;
     public String articleNumber;
-    public Double price;
+    public BigDecimal price;
     public Double availableStock;
     public String category;
