+++ src/main/java/com/example/sparkle/dtos/outputDto/ProductOutputDto.java
@@ -1,10 +1,12 @@
 package com.example.sparkle.dtos.outputDto;
 
+import java.math.BigDecimal;
+
 public class ProductOutputDto {
     //    DTO Variables
     public Long id;
     public String productName;
-    public Double price;
+    public BigDecimal price;
     public Double availableStock;
 
 }
